+++ Sources/SF2Lib/include/SF2Lib/Types.hpp
@@ -35,7 +35,7 @@ struct Accelerated
   using ConversionProc = void (*)(const int16_t*, vDSP_Stride, T*, vDSP_Stride, vDSP_Length);
   inline static ConversionProc conversionProc = []() noexcept {
     if constexpr (std::is_same_v<T, float>) return vDSP_vflt16;
-    if constexpr (std::is_same_v<T, Float>) return vDSP_vflt16D;
+    if constexpr (std::is_same_v<T, double>) return vDSP_vflt16D;
   }();
 
   /**
@@ -45,7 +45,7 @@ struct Accelerated
   using ScaleProc = void (*)(const T*, vDSP_Stride, const T*, T*, vDSP_Stride, vDSP_Length);
   inline static ScaleProc scaleProc = []() noexcept {
     if constexpr (std::is_same_v<T, float>) return vDSP_vsdiv;
-    if constexpr (std::is_same_v<T, Float>) return vDSP_vsdivD;
+    if constexpr (std::is_same_v<T, double>) return vDSP_vsdivD;
   }();
 
   /**
@@ -55,7 +55,7 @@ struct Accelerated
   using MagnitudeProc = void (*)(const T*, vDSP_Stride, T*, vDSP_Length);
   inline static MagnitudeProc magnitudeProc = []() noexcept {
     if constexpr (std::is_same_v<T, float>) return vDSP_maxmgv;
-    if constexpr (std::is_same_v<T, Float>) return vDSP_maxmgvD;
+    if constexpr (std::is_same_v<T, double>) return vDSP_maxmgvD;
   }();
 };
